Add redirectUri and scope to state sent to the authorization server and forwarded to the redirectUri.
I'm receiving a token now from my test token server. Just need to redirect back to the redirectBackUri, with that token and the original state encoded.

I will also move Config.elm into a .json file and probe every two seconds for changes, so you don't have to stop, rebuild, and restart just to make changes, and so that you don't need an Elm compiler on the server.