OAuthMiddleware.getAuthorization and getAuthorizations now take a useCache parameter.
Move the apiUri from the example to the apiUri field of EncodeDecode.Authorization.

A 'comment' field in the Authorization JSON now comments out that whole object.

Add real data to 'example/authorizations.json.template' for the four providers I've tested. User just has to copy to 'authorizations.json', uncomment, and fill in clientId and redirectUri.